+++ src/toisto/ui/text.py
@@ -9,6 +9,7 @@ from rich.console import Console
 from rich.panel import Panel
 
 from ..metadata import CHANGELOG_URL, NAME, README_URL, VERSION
+from ..model.language.iana_language_subtag_registry import ALL_LANGUAGES
 from ..model.language.label import END_OF_SENTENCE_PUNCTUATION, Label
 from ..model.quiz.quiz import Quiz
 from .dictionary import DICTIONARY_URL, linkify_and_enumerate
@@ -63,7 +64,7 @@ INCORRECT: Final = "❌ Incorrect. "
 
 def feedback_correct(guess: Label, quiz: Quiz) -> str:
     """Return the feedback about a correct result."""
-    return CORRECT + meaning(quiz) + other_answers(guess, quiz) + answer_notes(quiz) + examples(quiz)
+    return CORRECT + colloquial(quiz) + meaning(quiz) + other_answers(guess, quiz) + answer_notes(quiz) + examples(quiz)
 
 
 def feedback_incorrect(guess: Label, quiz: Quiz) -> str:
@@ -71,11 +72,19 @@ def feedback_incorrect(guess: Label, quiz: Quiz) -> str:
     if guess == Label(quiz.answer_language, "?"):
         answers = quiz.non_generated_answers
         label = "The correct answer is" if len(answers) == 1 else "The correct answers are"
-        feedback = f"{label} {linkify_and_enumerate(*answers)}.\n" + meaning(quiz)
+        feedback = f"{label} {linkify_and_enumerate(*answers)}.\n" + colloquial(quiz) + meaning(quiz)
     else:
         label = f'{INCORRECT}The correct answer is "{colored_diff(guess, quiz.answer)}".\n'
-        feedback = label + meaning(quiz) + other_answers(quiz.answer, quiz)
+        feedback = label + colloquial(quiz) + meaning(quiz) + other_answers(quiz.answer, quiz)
     return feedback + answer_notes(quiz)
+
+
+def colloquial(quiz: Quiz) -> str:
+    """Return the feedback about colloquial label, if any."""
+    if not quiz.question.is_colloquial:
+        return ""
+    feedback = f'The colloquial {ALL_LANGUAGES[quiz.question_language]} spoken was "{quiz.question.strip("*")}".'
+    return wrap(feedback, SECONDARY)
 
 
 def meaning(quiz: Quiz) -> str:
@@ -86,14 +95,14 @@ def meaning(quiz: Quiz) -> str:
         meanings = f"{question_meanings}, respectively {answer_meanings}"
     else:
         meanings = linkify_and_enumerate(*(quiz.question_meanings + quiz.answer_meanings))
-    return f"[{SECONDARY}]Meaning {meanings}.[/{SECONDARY}]\n" if meanings else ""
+    return wrap(f"Meaning {meanings}.", SECONDARY) if meanings else ""
 
 
 def other_answers(guess: Label, quiz: Quiz) -> str:
     """Return the quiz's other answers, if any."""
     if answers := quiz.other_answers(guess):
         label = "Another correct answer is" if len(answers) == 1 else "Other correct answers are"
-        return f"""[{SECONDARY}]{label} {linkify_and_enumerate(*answers)}.[/{SECONDARY}]\n"""
+        return wrap(f"{label} {linkify_and_enumerate(*answers)}.", SECONDARY)
     return ""
 
 
@@ -113,7 +122,7 @@ def answer_notes(quiz: Quiz) -> str:
 
 def instruction(quiz: Quiz) -> str:
     """Return the instruction for the quiz."""
-    return f"[{QUIZ}]{quiz.instruction}:[/{QUIZ}]"
+    return wrap(f"{quiz.instruction}:", QUIZ, postfix="")
 
 
 def show_welcome(write_output: Callable[..., None], latest_version: str | None, config: ConfigParser) -> None:
@@ -133,5 +142,10 @@ def bulleted_list(label: str, items: Sequence[str], style: str = SECONDARY, bull
     if len(items) == 0:
         return ""
     if len(items) == 1:
-        return f"[{style}]{label}: {items[0]}[/{style}]\n"
-    return f"[{style}]{label}s:\n" + "\n".join([f"{bullet} {item}" for item in items]) + f"[/{style}]\n"
+        return wrap(f"{label}: {items[0]}", style)
+    return wrap(f"{label}s:\n" + "\n".join([f"{bullet} {item}" for item in items]), style)
+
+
+def wrap(text: str, style: str, postfix: str = "\n") -> str:
+    """Wrap the text with the style."""
+    return f"[{style}]{text}[/{style}]{postfix}"